# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water track",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "Application for monitoring water consumption"
  },
  "servers": [
    {
      "url": "https://warettrack.onrender.com"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/water": {
      "post": {
        "tags": [
          "Water"
        ],
        "summary": "Create a new water intake record",
        "operationId": "createWater",
        "description": "This endpoint allows a user to create a new water intake record, specifying the amount, the date of intake, and the current daily water norm. A valid user ID is required.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Water intake data to be created",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "date",
                  "currentDailyNorm"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "example": 250,
                    "description": "The amount of water consumed in milliliters.",
                    "minimum": 50
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-11-26T15:30:45",
                    "description": "The date and time when the water was consumed.",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{3})?Z$"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a water record!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a water record!"
                    },
                    "data": {
                      "type": "array",
                      "description": "creating a water record.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "674c3649bc0f53cead7b8351",
                            "description": "Water id (must be a valid ObjectId)."
                          },
                          "date": {
                            "type": "string",
                            "description": "The date for which water records are aggregated.",
                            "example": "2024-11-26T15:30:45"
                          },
                          "amount": {
                            "type": "number",
                            "description": "Total amount of water consumed on the specific day in milliliters.",
                            "example": 2500
                          },
                          "currentDailyNorm": {
                            "type": "number",
                            "example": 2000
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid or missing parameters in the request body",
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid or expired access token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Something went wrong on the server side.",
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/api/water/{waterId}": {
      "patch": {
        "tags": [
          "Water"
        ],
        "summary": "Update a water intake record by ID",
        "operationId": "updateWaterById",
        "description": "This endpoint allows you to update an existing water intake record based on the provided water ID. A valid water ID is required, and you must provide the data to update.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "waterId",
            "required": true,
            "description": "The unique identifier of the water intake record (ObjectId).",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "example": "60c72b2f9b1d8e6b6d2e5f53"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "example": 50,
                    "description": "The amount of water consumed in milliliters(optional)."
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-11-26T15:30:45",
                    "description": "The date and time when the water was consumed(optional)."
                  },
                  "currentDailyNorm": {
                    "type": "number",
                    "example": 2000,
                    "description": "The current daily water intake norm(optional)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the water intake record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated the water record!"
                    },
                    "data": {
                      "type": "array",
                      "description": "response after data update",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "674c3649bc0f53cead7b8351",
                            "description": "Water id (must be a valid ObjectId)."
                          },
                          "date": {
                            "type": "string",
                            "description": "The date for which water records are aggregated.",
                            "example": "2024-11-26T15:30:45"
                          },
                          "amount": {
                            "type": "number",
                            "description": "Total amount of water consumed on the specific day in milliliters.",
                            "example": 2500
                          },
                          "currentDailyNorm": {
                            "type": "number",
                            "example": 2000
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid or missing parameters in the request body",
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid or expired access token"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Water record not found",
            "$ref": "#/components/responses/404"
          },
          "500": {
            "description": "Internal Server Error - Something went wrong on the server side.",
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Water"
        ],
        "summary": "Delete a water intake record by ID",
        "operationId": "deleteWaterById",
        "description": "This endpoint allows you to delete a water intake record based on the provided water ID. A valid water ID is required.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "waterId",
            "required": true,
            "description": "The unique identifier of the water intake record (ObjectId).",
            "schema": {
              "type": "string",
              "example": "60c72b2f9b1d8e6b6d2e5f53"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted the water intake record"
          },
          "400": {
            "description": "Bad Request - Invalid water ID format",
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid or expired access token"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Water record not found",
            "$ref": "#/components/responses/404"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/api/water/day/{date}": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Retrieve daily water intake records",
        "operationId": "getWaterPerDay",
        "description": "This endpoint allows you to retrieve all water intake records for a specific day. A valid date and authentication token are required.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "description": "The date for which to retrieve water intake records (ISO format, e.g., 2024-11-29).",
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2024-11-29"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved daily water records!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully retrieved daily water records!"
                    },
                    "data": {
                      "type": "array",
                      "description": "received water data for a month .",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "674c3649bc0f53cead7b8351",
                            "description": "Water id (must be a valid ObjectId)."
                          },
                          "date": {
                            "type": "string",
                            "description": "The date for which water records are aggregated.",
                            "example": "2024-11-26T15:30:45"
                          },
                          "amount": {
                            "type": "number",
                            "description": "Total amount of water consumed on the specific day in milliliters.",
                            "example": 2500
                          },
                          "currentDailyNorm": {
                            "type": "number",
                            "example": 2000
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid or missing parameters in the request.",
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid or expired access token"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - No water records found for the specified date.",
            "$ref": "#/components/responses/404"
          },
          "500": {
            "description": "Internal Server Error - Something went wrong on the server side.",
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/api/water/month/{date}": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Retrieve monthly water intake records",
        "operationId": "getWaterPerMonth",
        "description": "This endpoint allows you to retrieve all water intake records for a specific month. A valid date (ISO format) and authentication token are required.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "required": true,
            "description": "The date representing the month to retrieve water intake records for (ISO format, e.g., 2024-11-01).",
            "schema": {
              "type": "string",
              "format": "date",
              "example": "2024-11-01"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved monthly water records!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully retrieved monthly water records!"
                    },
                    "data": {
                      "type": "array",
                      "description": "received water data for a month.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "674c3649bc0f53cead7b8351",
                            "description": "Water id (must be a valid ObjectId)."
                          },
                          "date": {
                            "type": "string",
                            "description": "The date for which water records are aggregated.",
                            "example": "2024-11-26T15:30:45"
                          },
                          "amount": {
                            "type": "number",
                            "description": "Total amount of water consumed on the specific day in milliliters.",
                            "example": 2500
                          },
                          "currentDailyNorm": {
                            "type": "number",
                            "example": 2000
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid or missing parameters in the request.",
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Invalid or expired access token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 401
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid or expired access token"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found - No water records found for the specified month.",
            "$ref": "#/components/responses/404"
          },
          "500": {
            "description": "Internal Server Error - Something went wrong on the server side.",
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request - Invalid or missing parameters",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Bad request, invalid or missing parameters"
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Element not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Element not found"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal server error"
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}